# tunning hasSlot function and fix store wrong value in usedSlots