@avagin 2021-07-27
> BTW, in most cases, usedSlots[] is read rather than writen, and it stored unsorted physical addresses. So the function hasSlot is relatively less effective. Is it better to use hashmap to store the mapped physical address, thus we can have O(1) when search mapped physical addresses?

Adding a new element to a hashmap can require to do memory allocations. Maybe it was a reason why hashmap isn't used here.